Match fn_802B19F8 (95.1->100%)
(u16) casts on the getButtons* returns emit the clrlwi-before-sth the target
uses for the u16 field stores (#58); == 0u forces cmplwi on the masked flag
test (#3).

(cherry picked from commit 41f85edf4c1d62c92ff25e9884d9574f783366fc)

+++ src/main/dll/player.c
@@ -10918,15 +10918,15 @@ void fn_802B19F8(int obj, int state, f32 fv)
     ((PlayerState*)state)->buttonsHeld = 0;
     ((PlayerState*)state)->buttonsJustPressed = 0;
     ((PlayerState*)state)->buttonsJustPressedIfNotBusy = 0;
-    if ((((PlayerState*)state)->flags360 & 0x200000) == 0 &&
+    if ((((PlayerState*)state)->flags360 & 0x200000) == 0u &&
         ((PlayerState*)state)->characterId != -1 &&
         (c = ((PlayerState*)state)->curAnimId) != 0x44 && c != 0x4e)
     {
         *(int*)((char*)state + 0x6d0) = padGetStickX(0);
         *(int*)((char*)state + 0x6d4) = padGetStickY(0);
-        ((PlayerState*)state)->buttonsHeld = getButtonsHeld(0);
-        ((PlayerState*)state)->buttonsJustPressed = getButtonsJustPressed(0);
-        ((PlayerState*)state)->buttonsJustPressedIfNotBusy = getButtonsJustPressedIfNotBusy(0);
+        ((PlayerState*)state)->buttonsHeld = (u16)getButtonsHeld(0);
+        ((PlayerState*)state)->buttonsJustPressed = (u16)getButtonsJustPressed(0);
+        ((PlayerState*)state)->buttonsJustPressedIfNotBusy = (u16)getButtonsJustPressedIfNotBusy(0);
     }
     *(f32*)((char*)state + 0x6dc) = (f32) * (int*)((char*)state + 0x6d0);
     *(f32*)((char*)state + 0x6d8) = (f32) * (int*)((char*)state + 0x6d4);
